+++ CountlyConfig.h
@@ -529,6 +529,13 @@ typedef enum : NSUInteger
  * @discussion For disabling it, please set this flag to @c NO.
  */
 @property (nonatomic) BOOL enableOrientationTracking;
+
+/**
+ * This is an experimental feature
+ * For enabling fetching and application of server config values.
+ * @discussion If set, Server Config values from Countly Server will be fetched at the beginning of a session.
+ */
+@property (nonatomic) BOOL enableServerConfiguration;
 NS_ASSUME_NONNULL_END
 
 @end