+++ CountlyConfig.m
@@ -49,7 +49,7 @@ CLYDeviceIDType const CLYDeviceIDTypeNSUUID     = @"CLYDeviceIDTypeNSUUID";
 
         self.maxKeyLength = 128;
         self.maxValueLength = 256;
-        self.maxSegmentationValues = 30;
+        self.maxSegmentationValues = 100;
 
         self.location = kCLLocationCoordinate2DInvalid;
 
@@ -58,6 +58,7 @@ CLYDeviceIDType const CLYDeviceIDTypeNSUUID     = @"CLYDeviceIDTypeNSUUID";
         self.internalLogLevel = CLYInternalLogLevelDebug;
 
         self.enableOrientationTracking = YES;
+        self.enableServerConfiguration = NO;
     }
 
     return self;