+++ CountlyEvent.h
@@ -12,6 +12,7 @@
 @property (nonatomic, copy) NSString* ID;
 @property (nonatomic, copy) NSString* CVID;
 @property (nonatomic, copy) NSString* PVID;
+@property (nonatomic, copy) NSString* PEID;
 @property (nonatomic, copy) NSDictionary* segmentation;
 @property (nonatomic) NSUInteger count;
 @property (nonatomic) double sum;